Get to know Web Workers sync status in UI (#2754)
# Motivation

It will be useful in the future, notably for #2639, to display an
information in the UI to the users about the status of the Web Workers
synchronization. Notably displaying a small hint like a spinner when
work is in progress.

This PR introduces a store that contains such information for upcoming
workers of #2639 and also improve the timer utils to transmit the
information

# Changes

- new `sync.store` and `sync.derived`
- add `postMessage` in timer utils to transmit the sync status to the UI