More details in merge neurons simulation (#2802)
# Motivation

We want to show the results or merging neurons before the user commits.
This PR adds more details in the `ConfirmNeuronsMerge` steps.
Here you can see what it looks like but the details are not important
because this is behind a feature flag just to try it out and iterate on:

<img width="611" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/84eb3d97-09fe-4590-ae27-f318ca373108">


# Changes

1. Add new component `NnsNeuronDetailCard`
2. Use it instead of `NnsNeuronInfo` in `ConfirmNeuronsMerge`

# Tests

1. `frontend/src/tests/lib/modals/neurons/MergeNeuronsModal.spec.ts` is
modified to work with the new card.
2. A new test is added for the new component.
3. A page object is added for the new component.

# Todos

- [x] Add entry to changelog (if necessary).