GIX-1658: Fix task runner substitutes canister (#2806)
# Motivation

User wants to add a name when creating a canister.

There was a bug in the create canister flow.

When creating a canister, the nns-dapp backend has a runner task that
attached a newly created canister. This method was overriding the
previous attched canister by the method `attach_canister` with an empty
name.

# Changes

* Attach the canister in the runner flow only if canister_id is not
alreay attached.
* Manage race condition in `attach_canister` method. The runner could be
faster than the call to this request.

# Tests

* Add tests for `attach_newly_created_canister`.
* Add test for new condition in `attach_canister`.

# Todos

- [x] Add entry to changelog (if necessary).