Observe ckBTC balances (#2810)
# Motivation

Use the new web workers to observe the ckBTC balances in the dedicated
accounts and wallet page.

# Notes

In the wallet, to refresh the balance, we have to reload the account
from store. Because fetching the transactions starts when the account
changes, the dapp end up doing an unnecessary call to for the
transactions. Something we can try to improve in another PR.

# Changes

- new `CkBTCBalancesObserver` component
- use above component in `CkBTCAccounts` and `CkBTCWallet`

Not related: I also renamed the test `SnsWalletBalancesObserver` to
`SnsBalancesObserver`, felt like we can spare a PR.

# Screenshot


![demo2](https://github.com/dfinity/nns-dapp/assets/16886711/91dec897-7d35-4e3b-8410-327c8313e9b4)