Call result patches (#2826)
# Motivation
The patch files applied when generating Rust from SNS .did files are
quite large and break easily. There are a few changes in the automatic
generation that can make the patch files significantly smaller. Here is
the first:

Service definitions have return type `CallResult` but `didc` generates
`Result` in the rust code, with the expectation that the `CallResult`
will be imported as `Result`. Unfortunately that expectation can often
not be met as `Result` is used for other purposes. So the patch files
have many lines dedicated to changeing the `Result` in the API
signatures to `CallResult`. This can be automated.

# Changes
- Use `CallResult` instead of `Result` when auto-generating code.
- Delete patch file lines that used to do the same.

# Tests
Observe that the test "IC Commit" generates the rust files and gets the
same result as before.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>