Make sure aggregator data collection happens during the test (#2840)
# Motivation

The newest snsdemo snapshot (which we don't use yet), has the
sns_aggregator installed and SNSes preloaded.
We have a test that tests that data is collected by checking that there
are 10 SNSes on the first page.
But we don't (yet) check that those SNSes were loaded during the test.

# Changes

Add a step in the workflow where we check that the number of loaded
SNSes is < 10 to make sure that data is actually collected during the
test.

# Tests

yes

# Todos

This is in preparation to use the new snapshot which has the aggregator
preloaded.
I'll add an entry when for that PR.
- [ ] Add entry to changelog (if necessary).