Empty record for enums as well (#2844)
# Motivation
The patch files applied when generating Rust from SNS .did files have
grown quite large and break easily. There are a few changes in the
automatic generation that can make the patch files significantly
smaller. Here is the fifth:

# Changes
* `EmptyRecord` is (almost always) used in enum entries without data.
That can be automated.
* Note: Interestingly, one empty enum entry has no `EmptyRecord`. It
would be interesting to know why. Is the whole `EmptyRecord` dance not
needed any more, or at least, not in enums? Or is that a special case
that works for some reason? However that is not the focus of this PR.

# Tests
* Proof of correctness: Note that the Rust code is unchanged, so the sns
aggregator will behave in the same way. The patch files have dropped
from 339 lines to 266.

# Todos

- [x] Add entry to changelog (if necessary).
There is already an entry for changes to `EmptyRecord` to reduce the
size of the patch file.

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>