Revert display of TVL in various currencies to USD only (#2853)
# Motivation

We were requested to revert the display of TVL in various currencies and
display again only USD only.

# Note

In case this would change again in the future, this PR only removes the
randomization of the parameters in the query and keep all the logic in
place.

PR reverts what was introduced in #2695.

# Changes

- remove random currency parameter in TVL request (undefined = dollar
per default)

# Todos

- [ ] Add entry to changelog (if necessary).