Update ic-cdk (#2856)
# Motivation
Actually, the goal was to update candid to the latest version. It seems
that this is not possible yet, but it is possible to update the `ic-cdk`
so we might as well do that while the candid issue is resolved.

# Changes
* Update `ic-cdk` to the latest version in the aggregator.
* Use `ic-cdk-timers` as that has been split out from `ic-cdk`.

Note: The nns-dapp does not yet make meaningful use of the CDK, so that
is untouched in this PR.

# Tests
* See CI.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>