GIX-1725: Merge Only Locked Neurons (#2874)
# Motivation

As explained
[here](https://forum.dfinity.org/t/52-year-neuron-fixes/21301), only
neurons in the Locked state are mergeable until further notice.

To avoid confusion with the users, we should apply this condition on top
of the rest of conditions we use to disable neurons when merging them.

# Changes

* Add a new case in neuron util helper `isMergeableNeuron`.
* Add the same case in neuron util helper
`getMergeableNeuronMessageKey`.

# Tests

* Add a new teste case in `mapMergeableNeurons` which uses the helpers.
* Add the Locked state explicitly in the merging neurons tests. This was
not strictly necessary because the mock neuron is already in the locked
state. But I preferred to be explicit that the locked state is required
for a mergeable neuron.

# Todos

- [x] Add entry to changelog (if necessary).