Prefix accounts.store with icp (#2897)
# Motivation

Now that we have various tokens and dedicated stores
(`icrc-accounts.store`), I was thinking that prefixing `accounts.store`
with `icp` - i.e. `icp-accounts.store` - makes it more explicit and
easier to read.

WDYT @lmuntaner ?

# Changes

- refactoring only, no functional changes

# Notes

There is another store and service that can be prefixed later as well in
separate PRs.