Display swaps lifecycle on aggregator landing page (#2904)
# Motivation

The dashboard used to not display the status of the swap and some used
the opportunity to took screenshots of the Snses to claim they were
active. By displaying the status, we prevent the same thing to happen
with the landing page of the aggregator.

# Changes

- display swap status on aggregator landing page

# Screenshot

<img width="1535" alt="Capture d’écran 2023-07-18 à 13 17 23"
src="https://github.com/dfinity/nns-dapp/assets/16886711/525dcc84-0c6b-477e-aece-6ac9dae83e15">