E2E: Always use 6 workers (#2912)
# Motivation

We are now using faster machines to run e2e tests on CI.
These should also be capable of running more tests in parallel.

# Changes

Always use 6 workers for e2e tests.

# Tests

Looking at the CI run, the tests are noticeably slower but not twice as
slow so in total it's a win.

# Todos

- [ ] Add entry to changelog (if necessary).

Not note-worthy.